:recycle: Separate files for Algorithm func
